backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
python-dotenv==1.0.1
pydantic>=2.9.0
pydantic-settings>=2.5.0
httpx==0.27.0
langchain>=0.3.7
langchain-openai>=0.2.5
langgraph>=0.2.28
elevenlabs>=1.9.0
websockets>=13.0
